Add a hash of matching opening / closing tokens
Instead of using substring, substitutions and conversion from/to symbols
for finding the opening token name of a given closing token, introduce a
hash of matching names.

This is not that useful for the existing tokens which have very similar
opening / closing names, but will make the code simpler for supporting
heredoc strings.

No functional change.